requirements1.txt
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
dropbox==12.0.2
ffmpeg==1.4
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
idna==3.7
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai-whisper==20231117
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
ply==3.11
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
regex==2024.7.24
requests==2.32.3
selenium==4.22.0
setuptools==71.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stone==3.3.1
sympy==1.13.1
tiktoken==0.7.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.1
websocket-client==1.8.0
wsproto==1.2.0
